Fix errors with mutation string literals
In ruby 2.3, an option can be set to freeze string literals in order to
improve performance. In order to comply with that change, all string
mutations were removed from the codebase.

Instead of extending a string with CoreExtensions::Object::Blank,
a new CoreExtensions::String is created with the target string literal
that encapsulates the desired behavior.

- add: Rescuetime::CoreExtensions::String plus unit tests