Bump to v1.0.0!
While there were some more features I wanted to implement before v1.0.0,
I'm not going to have a lot of time for feature development on this
project for the next couple of months. I'm pretty confident in the API
and anticipate no more breaking changes, however, so I'm just going to go
ahead and bump this to v1.0.0–while an incomplete implementation, it is
ready for use.