Update the Code of Conduct
Update the project to use a more recent version of the Contributor Covenant
as a base. It's been around for a while now and has been used by many large
open source projects, so the current version is much more "battle-tested"
than v1.0.0.

In particular, I like the examples listed in "Our Standards" as well as the
"Enforcement" section additions.