Add TODO section to README
This adds a TODO section (including completed development milestones)
to the end of the README in the Contributing section. Basically, it's just
copying the Development Map from the wiki.

My hope is that by putting the TODO items front-and-center, it will 1) give
potential gem users a better idea of where this library is going (and more
security about using it), and 2) help potential contributors understand
development priorities.

I will need to open more issues and add them in later on.